chat bot and order service added in ecom system

+++ server_spring/src/main/java/in/heuton/skipper/entity/dto/OrderItemRequest.java
@@ -1,7 +1,7 @@
 package in.heuton.skipper.entity.dto;
 
 public record OrderItemRequest(
-        int productId,
+        long productId,
         int quantity
 )
 {
